refactor: split files and make more functions inlined
