+++ src/treespec/treespec.cpp
@@ -17,57 +17,23 @@ limitations under the License.
 
 #include "include/treespec.h"
 
-#include <algorithm>      // std::copy, std::reverse
-#include <exception>      // std::rethrow_exception, std::current_exception
-#include <iterator>       // std::back_inserter
-#include <memory>         // std::unique_ptr, std::make_unique
-#include <optional>       // std::optional
-#include <sstream>        // std::ostringstream
-#include <stdexcept>      // std::runtime_error
-#include <string>         // std::string
-#include <thread>         // std::thread::id, std::this_thread
-#include <tuple>          // std::tuple
-#include <unordered_map>  // std::unordered_map
-#include <unordered_set>  // std::unordered_set
-#include <utility>        // std::move, std::pair
-#include <vector>         // std::vector
+#include <algorithm>  // std::copy, std::reverse
+#include <iterator>   // std::back_inserter
+#include <memory>     // std::unique_ptr, std::make_unique
+#include <optional>   // std::optional
+#include <sstream>    // std::ostringstream
+#include <tuple>      // std::tuple
+#include <utility>    // std::move
+#include <vector>     // std::vector
 
 #include "include/critical_section.h"
 #include "include/exceptions.h"
-#include "include/mutex.h"
+#include "include/hashing.h"
+#include "include/pytypes.h"
 #include "include/registry.h"
-#include "include/utils.h"
+#include "include/stdutils.h"
 
 namespace optree {
-
-/*static*/ std::string PyTreeSpec::NodeKindToString(const Node& node) {
-    switch (node.kind) {
-        case PyTreeKind::Leaf:
-            return "leaf type";
-        case PyTreeKind::None:
-            return "NoneType";
-        case PyTreeKind::Tuple:
-            return "tuple";
-        case PyTreeKind::List:
-            return "list";
-        case PyTreeKind::Dict:
-            return "dict";
-        case PyTreeKind::OrderedDict:
-            return "OrderedDict";
-        case PyTreeKind::DefaultDict:
-            return "defaultdict";
-        case PyTreeKind::NamedTuple:
-        case PyTreeKind::StructSequence:
-            return PyRepr(node.node_data);
-        case PyTreeKind::Deque:
-            return "deque";
-        case PyTreeKind::Custom:
-            EXPECT_NE(node.custom, nullptr, "The custom registration is null.");
-            return PyRepr(node.custom->type);
-        default:
-            INTERNAL_ERROR();
-    }
-}
 
 // NOLINTNEXTLINE[readability-function-cognitive-complexity]
 /*static*/ py::object PyTreeSpec::MakeNode(const Node& node,
@@ -265,14 +231,15 @@ namespace optree {
         return {pos - cur, other_pos - other_cur, root.num_nodes, root.num_leaves};
     }
 
-    Node node;
-    node.kind = root.kind;
-    node.arity = root.arity;
-    node.num_leaves = 0;
-    node.num_nodes = 1;
-    node.node_data = root.node_data;
-    node.custom = root.custom;
-    node.original_keys = root.original_keys;
+    Node node{
+        .kind = root.kind,
+        .arity = root.arity,
+        .node_data = root.node_data,
+        .custom = root.custom,
+        .num_leaves = 0,
+        .num_nodes = 1,
+        .original_keys = root.original_keys,
+    };
     switch (root.kind) {
         case PyTreeKind::Tuple:
         case PyTreeKind::List:
@@ -857,22 +824,6 @@ std::unique_ptr<PyTreeSpec> PyTreeSpec::Child(ssize_t index) const {
     return child;
 }
 
-ssize_t PyTreeSpec::GetNumLeaves() const {
-    EXPECT_FALSE(m_traversal.empty(), "The tree node traversal is empty.");
-    return m_traversal.back().num_leaves;
-}
-
-ssize_t PyTreeSpec::GetNumNodes() const { return py::ssize_t_cast(m_traversal.size()); }
-
-ssize_t PyTreeSpec::GetNumChildren() const {
-    EXPECT_FALSE(m_traversal.empty(), "The tree node traversal is empty.");
-    return m_traversal.back().arity;
-}
-
-bool PyTreeSpec::GetNoneIsLeaf() const { return m_none_is_leaf; }
-
-std::string PyTreeSpec::GetNamespace() const { return m_namespace; }
-
 py::object PyTreeSpec::GetType(const std::optional<Node>& node) const {
     if (!node.has_value()) [[likely]] {
         EXPECT_FALSE(m_traversal.empty(), "The tree node traversal is empty.");
@@ -904,681 +855,6 @@ py::object PyTreeSpec::GetType(const std::optional<Node>& node) const {
         default:
             INTERNAL_ERROR();
     }
-}
-
-PyTreeKind PyTreeSpec::GetPyTreeKind() const {
-    EXPECT_FALSE(m_traversal.empty(), "The tree node traversal is empty.");
-    return m_traversal.back().kind;
-}
-
-bool PyTreeSpec::IsLeaf(const bool& strict) const {
-    if (strict) [[likely]] {
-        return GetNumNodes() == 1 && GetNumLeaves() == 1;
-    }
-    return GetNumNodes() == 1;
-}
-
-// NOLINTNEXTLINE[readability-function-cognitive-complexity]
-bool PyTreeSpec::IsPrefix(const PyTreeSpec& other, const bool& strict) const {
-    if (m_none_is_leaf != other.m_none_is_leaf) [[unlikely]] {
-        return false;
-    }
-    if (!m_namespace.empty() && !other.m_namespace.empty() && m_namespace != other.m_namespace)
-        [[likely]] {
-        return false;
-    }
-    if (GetNumNodes() > other.GetNumNodes()) [[likely]] {
-        return false;
-    }
-
-    bool all_leaves_match = true;
-    std::vector<Node> other_traversal{other.m_traversal};
-    // NOLINTNEXTLINE[readability-qualified-auto]
-    auto b = other_traversal.rbegin();
-    // NOLINTNEXTLINE[readability-qualified-auto]
-    for (auto a = m_traversal.rbegin(); a != m_traversal.rend(); ++a, ++b) {
-        if (b == other_traversal.rend()) [[unlikely]] {
-            return false;
-        }
-        if (a->kind == PyTreeKind::Leaf) [[unlikely]] {
-            all_leaves_match &= b->kind == PyTreeKind::Leaf;
-            b += b->num_nodes - 1;
-            EXPECT_LT(b, other_traversal.rend(), "PyTreeSpec traversal out of range.");
-            continue;
-        }
-        if (a->arity != b->arity ||
-            static_cast<bool>(a->node_data) != static_cast<bool>(b->node_data) ||
-            a->custom != b->custom) [[likely]] {
-            return false;
-        }
-
-        switch (a->kind) {
-            case PyTreeKind::None:
-            case PyTreeKind::Tuple:
-            case PyTreeKind::List:
-            case PyTreeKind::Deque: {
-                if (a->kind != b->kind) [[likely]] {
-                    return false;
-                }
-                break;
-            }
-
-            case PyTreeKind::Dict:
-            case PyTreeKind::OrderedDict:
-            case PyTreeKind::DefaultDict: {
-                if (b->kind != PyTreeKind::Dict && b->kind != PyTreeKind::OrderedDict &&
-                    b->kind != PyTreeKind::DefaultDict) [[likely]] {
-                    return false;
-                }
-                const scoped_critical_section2 cs(a->node_data, b->node_data);
-                const auto expected_keys = (a->kind != PyTreeKind::DefaultDict
-                                                ? py::reinterpret_borrow<py::list>(a->node_data)
-                                                : TupleGetItemAs<py::list>(a->node_data, 1));
-                const auto other_keys = (b->kind != PyTreeKind::DefaultDict
-                                             ? py::reinterpret_borrow<py::list>(b->node_data)
-                                             : TupleGetItemAs<py::list>(b->node_data, 1));
-                const py::dict dict{};
-                for (ssize_t i = 0; i < b->arity; ++i) {
-                    DictSetItem(dict, ListGetItem(other_keys, i), py::int_(i));
-                }
-                if (!DictKeysEqual(expected_keys, dict)) [[likely]] {
-                    return false;
-                }
-                if (expected_keys.not_equal(other_keys)) [[unlikely]] {
-                    auto other_offsets = reserved_vector<ssize_t>(b->arity + 1);
-                    auto other_num_nodes = reserved_vector<ssize_t>(b->arity);
-                    auto other_cur = b + 1;
-                    other_offsets.emplace_back(1);
-                    for (ssize_t j = b->arity - 1; j >= 0; --j) {
-                        const ssize_t num_nodes = other_cur->num_nodes;
-                        other_num_nodes.emplace_back(num_nodes);
-                        other_offsets.emplace_back(other_offsets.back() + num_nodes);
-                        other_cur += num_nodes;
-                    }
-                    std::reverse(other_num_nodes.begin(), other_num_nodes.end());
-                    std::reverse(other_offsets.begin(), other_offsets.end());
-                    EXPECT_EQ(other_offsets.front(),
-                              b->num_nodes,
-                              "PyTreeSpec traversal out of range.");
-                    auto reordered_index_to_index = std::unordered_map<ssize_t, ssize_t>{};
-                    for (ssize_t i = a->arity - 1; i >= 0; --i) {
-                        const py::object key = ListGetItem(expected_keys, i);
-                        reordered_index_to_index.emplace(i,
-                                                         py::cast<ssize_t>(DictGetItem(dict, key)));
-                    }
-                    auto reordered_other_num_nodes = reserved_vector<ssize_t>(b->arity);
-                    reordered_other_num_nodes.resize(b->arity);
-                    for (const auto& [i, j] : reordered_index_to_index) {
-                        reordered_other_num_nodes[i] = other_num_nodes[j];
-                    }
-                    auto reordered_other_offsets = reserved_vector<ssize_t>(b->arity + 1);
-                    reordered_other_offsets.emplace_back(1);
-                    for (ssize_t i = a->arity - 1; i >= 0; --i) {
-                        reordered_other_offsets.emplace_back(reordered_other_offsets.back() +
-                                                             reordered_other_num_nodes[i]);
-                    }
-                    std::reverse(reordered_other_offsets.begin(), reordered_other_offsets.end());
-                    EXPECT_EQ(reordered_other_offsets.front(),
-                              b->num_nodes,
-                              "PyTreeSpec traversal out of range.");
-                    auto original_b = other.m_traversal.rbegin() + (b - other_traversal.rbegin());
-                    for (const auto& [i, j] : reordered_index_to_index) {
-                        std::copy(original_b + other_offsets[j + 1],
-                                  original_b + other_offsets[j],
-                                  b + reordered_other_offsets[i + 1]);
-                    }
-                }
-                break;
-            }
-
-            case PyTreeKind::NamedTuple:
-            case PyTreeKind::StructSequence:
-            case PyTreeKind::Custom: {
-                const scoped_critical_section2 cs(a->node_data, b->node_data);
-                if (a->kind != b->kind || (a->node_data && a->node_data.not_equal(b->node_data)))
-                    [[likely]] {
-                    return false;
-                }
-                break;
-            }
-
-            case PyTreeKind::Leaf:
-            default:
-                INTERNAL_ERROR();
-        }
-
-        if (a->num_nodes > b->num_nodes) [[likely]] {
-            return false;
-        }
-    }
-    EXPECT_EQ(b, other_traversal.rend(), "PyTreeSpec traversal did not yield a singleton.");
-    return !strict || !all_leaves_match;
-}
-
-bool PyTreeSpec::operator==(const PyTreeSpec& other) const {
-    if (m_traversal.size() != other.m_traversal.size() || m_none_is_leaf != other.m_none_is_leaf)
-        [[likely]] {
-        return false;
-    }
-    if (!m_namespace.empty() && !other.m_namespace.empty() && m_namespace != other.m_namespace)
-        [[likely]] {
-        return false;
-    }
-
-    // NOLINTNEXTLINE[readability-qualified-auto]
-    auto b = other.m_traversal.cbegin();
-    // NOLINTNEXTLINE[readability-qualified-auto]
-    for (auto a = m_traversal.cbegin(); a != m_traversal.cend(); ++a, ++b) {
-        if (a->kind != b->kind || a->arity != b->arity ||
-            static_cast<bool>(a->node_data) != static_cast<bool>(b->node_data) ||
-            a->custom != b->custom) [[likely]] {
-            return false;
-        }
-        const scoped_critical_section2 cs(a->node_data, b->node_data);
-        if (a->node_data && a->node_data.not_equal(b->node_data)) [[likely]] {
-            return false;
-        }
-        EXPECT_EQ(a->num_leaves, b->num_leaves);
-        EXPECT_EQ(a->num_nodes, b->num_nodes);
-    }
-    return true;
-}
-
-// NOLINTNEXTLINE[readability-function-cognitive-complexity]
-std::string PyTreeSpec::ToStringImpl() const {
-    auto agenda = reserved_vector<std::string>(4);
-    for (const Node& node : m_traversal) {
-        EXPECT_GE(py::ssize_t_cast(agenda.size()), node.arity, "Too few elements for container.");
-
-        std::ostringstream children_sstream{};
-        {
-            bool first = true;
-            for (auto it = agenda.cend() - node.arity; it != agenda.cend(); ++it) {
-                if (!first) [[likely]] {
-                    children_sstream << ", ";
-                }
-                children_sstream << *it;
-                first = false;
-            }
-        }
-        const std::string children = children_sstream.str();
-
-        std::ostringstream sstream{};
-        switch (node.kind) {
-            case PyTreeKind::Leaf: {
-                sstream << "*";
-                break;
-            }
-
-            case PyTreeKind::None: {
-                sstream << "None";
-                break;
-            }
-
-            case PyTreeKind::Tuple: {
-                sstream << "(" << children;
-                // Tuples with only one element must have a trailing comma.
-                if (node.arity == 1) [[unlikely]] {
-                    sstream << ",";
-                }
-                sstream << ")";
-                break;
-            }
-
-            case PyTreeKind::List: {
-                sstream << "[" << children << "]";
-                break;
-            }
-
-            case PyTreeKind::Dict:
-            case PyTreeKind::OrderedDict: {
-                const scoped_critical_section cs{node.node_data};
-                EXPECT_EQ(ListGetSize(node.node_data),
-                          node.arity,
-                          "Number of keys and entries does not match.");
-                if (node.kind == PyTreeKind::OrderedDict) [[unlikely]] {
-                    sstream << "OrderedDict(";
-                }
-                if (node.kind == PyTreeKind::Dict || node.arity > 0) [[likely]] {
-                    sstream << "{";
-                }
-                bool first = true;
-                auto child_iter = agenda.cend() - node.arity;
-                for (const py::handle& key : node.node_data) {
-                    if (!first) [[likely]] {
-                        sstream << ", ";
-                    }
-                    sstream << PyRepr(key) << ": " << *child_iter;
-                    ++child_iter;
-                    first = false;
-                }
-                if (node.kind == PyTreeKind::Dict || node.arity > 0) [[likely]] {
-                    sstream << "}";
-                }
-                if (node.kind == PyTreeKind::OrderedDict) [[unlikely]] {
-                    sstream << ")";
-                }
-                break;
-            }
-
-            case PyTreeKind::NamedTuple: {
-                const py::object type = node.node_data;
-                const auto fields = NamedTupleGetFields(type);
-                EXPECT_EQ(TupleGetSize(fields),
-                          node.arity,
-                          "Number of fields and entries does not match.");
-                const std::string kind =
-                    PyStr(EVALUATE_WITH_LOCK_HELD(py::getattr(type, Py_Get_ID(__name__)), type));
-                sstream << kind << "(";
-                bool first = true;
-                auto child_iter = agenda.cend() - node.arity;
-                for (const py::handle& field : fields) {
-                    if (!first) [[likely]] {
-                        sstream << ", ";
-                    }
-                    sstream << PyStr(field) << "=" << *child_iter;
-                    ++child_iter;
-                    first = false;
-                }
-                sstream << ")";
-                break;
-            }
-
-            case PyTreeKind::DefaultDict: {
-                const scoped_critical_section cs(node.node_data);
-                EXPECT_EQ(TupleGetSize(node.node_data), 2, "Number of metadata mismatch.");
-                const py::object default_factory = TupleGetItem(node.node_data, 0);
-                const auto keys = TupleGetItemAs<py::list>(node.node_data, 1);
-                EXPECT_EQ(ListGetSize(keys),
-                          node.arity,
-                          "Number of keys and entries does not match.");
-                sstream << "defaultdict(" << PyRepr(default_factory) << ", {";
-                bool first = true;
-                auto child_it = agenda.cend() - node.arity;
-                for (const py::handle& key : keys) {
-                    if (!first) [[likely]] {
-                        sstream << ", ";
-                    }
-                    sstream << PyRepr(key) << ": " << *child_it;
-                    ++child_it;
-                    first = false;
-                }
-                sstream << "})";
-                break;
-            }
-
-            case PyTreeKind::Deque: {
-                sstream << "deque([" << children << "]";
-                if (!node.node_data.is_none()) [[unlikely]] {
-                    sstream << ", maxlen=" << PyRepr(node.node_data);
-                }
-                sstream << ")";
-                break;
-            }
-
-            case PyTreeKind::StructSequence: {
-                const py::object type = node.node_data;
-                const auto fields = StructSequenceGetFields(type);
-                EXPECT_EQ(TupleGetSize(fields),
-                          node.arity,
-                          "Number of fields and entries does not match.");
-                const py::object module_name = EVALUATE_WITH_LOCK_HELD(
-                    py::getattr(type, Py_Get_ID(__module__), Py_Get_ID(__main__)),
-                    type);
-                if (!module_name.is_none()) [[likely]] {
-                    const std::string name = PyStr(module_name);
-                    if (!(name.empty() || name == "__main__" || name == "builtins" ||
-                          name == "__builtins__")) [[likely]] {
-                        sstream << name << ".";
-                    }
-                }
-                const py::object qualname =
-                    EVALUATE_WITH_LOCK_HELD(py::getattr(type, Py_Get_ID(__qualname__)), type);
-                sstream << PyStr(qualname) << "(";
-                bool first = true;
-                auto child_iter = agenda.cend() - node.arity;
-                for (const py::handle& field : fields) {
-                    if (!first) [[likely]] {
-                        sstream << ", ";
-                    }
-                    sstream << PyStr(field) << "=" << *child_iter;
-                    ++child_iter;
-                    first = false;
-                }
-                sstream << ")";
-                break;
-            }
-
-            case PyTreeKind::Custom: {
-                const std::string kind = PyStr(
-                    EVALUATE_WITH_LOCK_HELD(py::getattr(node.custom->type, Py_Get_ID(__name__)),
-                                            node.custom->type));
-                sstream << "CustomTreeNode(" << kind << "[";
-                if (node.node_data) [[likely]] {
-                    sstream << PyRepr(node.node_data);
-                }
-                sstream << "], [" << children << "])";
-                break;
-            }
-
-            default:
-                INTERNAL_ERROR();
-        }
-
-        agenda.resize(agenda.size() - node.arity);
-        agenda.emplace_back(sstream.str());
-    }
-
-    EXPECT_EQ(agenda.size(), 1, "PyTreeSpec traversal did not yield a singleton.");
-    std::ostringstream oss{};
-    oss << "PyTreeSpec(" << agenda.back();
-    if (m_none_is_leaf) [[unlikely]] {
-        oss << ", NoneIsLeaf";
-    }
-    if (!m_namespace.empty()) [[unlikely]] {
-        oss << ", namespace=" << PyRepr(m_namespace);
-    }
-    oss << ")";
-    return oss.str();
-}
-
-std::string PyTreeSpec::ToString() const {
-    using ThreadIndent = std::pair<const PyTreeSpec*, std::thread::id>;
-    static std::unordered_set<ThreadIndent, ThreadIndentTypeHash> running{};
-    static read_write_mutex mutex{};
-
-    const ThreadIndent indent{this, std::this_thread::get_id()};
-    {
-        const scoped_read_lock_guard lock{mutex};
-        if (running.find(indent) != running.end()) [[unlikely]] {
-            return "...";
-        }
-    }
-
-    {
-        const scoped_write_lock_guard lock{mutex};
-        running.insert(indent);
-    }
-    try {
-        std::string representation = ToStringImpl();
-        {
-            const scoped_write_lock_guard lock{mutex};
-            running.erase(indent);
-        }
-        return representation;
-    } catch (...) {
-        {
-            const scoped_write_lock_guard lock{mutex};
-            running.erase(indent);
-        }
-        std::rethrow_exception(std::current_exception());
-    }
-}
-
-/*static*/ void PyTreeSpec::HashCombineNode(ssize_t& seed, const PyTreeSpec::Node& node) {
-    ssize_t data_hash = 0;
-    switch (node.kind) {
-        case PyTreeKind::Custom:
-            // We don't hash node_data custom node types since they may not hashable.
-            break;
-
-        case PyTreeKind::Leaf:
-        case PyTreeKind::None:
-        case PyTreeKind::Tuple:
-        case PyTreeKind::List:
-        case PyTreeKind::NamedTuple:
-        case PyTreeKind::Deque:
-        case PyTreeKind::StructSequence: {
-            const scoped_critical_section cs{node.node_data};
-            data_hash = py::hash(node.node_data ? node.node_data : py::none());
-            break;
-        }
-
-        case PyTreeKind::Dict:
-        case PyTreeKind::OrderedDict:
-        case PyTreeKind::DefaultDict: {
-            const scoped_critical_section cs{node.node_data};
-            if (node.kind == PyTreeKind::DefaultDict) [[unlikely]] {
-                EXPECT_EQ(TupleGetSize(node.node_data), 2, "Number of metadata mismatch.");
-                const py::object default_factory = TupleGetItem(node.node_data, 0);
-                data_hash = EVALUATE_WITH_LOCK_HELD(py::hash(default_factory), default_factory);
-            }
-            const auto keys = (node.kind != PyTreeKind::DefaultDict
-                                   ? py::reinterpret_borrow<py::list>(node.node_data)
-                                   : TupleGetItemAs<py::list>(node.node_data, 1));
-            EXPECT_EQ(ListGetSize(keys), node.arity, "Number of keys and entries does not match.");
-            for (const py::handle& key : keys) {
-                HashCombine(data_hash, EVALUATE_WITH_LOCK_HELD(py::hash(key), key));
-            }
-            break;
-        }
-
-        default:
-            INTERNAL_ERROR();
-    }
-
-    HashCombine(seed, node.kind);
-    HashCombine(seed, node.arity);
-    HashCombine(seed, node.custom);
-    HashCombine(seed, node.num_leaves);
-    HashCombine(seed, node.num_nodes);
-    HashCombine(seed, data_hash);
-}
-
-ssize_t PyTreeSpec::HashValueImpl() const {
-    ssize_t seed = 0;
-    for (const Node& node : m_traversal) {
-        HashCombineNode(seed, node);
-    }
-    HashCombine(seed, m_none_is_leaf);
-    HashCombine(seed, m_namespace);
-    return seed;
-}
-
-ssize_t PyTreeSpec::HashValue() const {
-    using ThreadIndent = std::pair<const PyTreeSpec*, std::thread::id>;
-    static std::unordered_set<ThreadIndent, ThreadIndentTypeHash> running{};
-    static read_write_mutex mutex{};
-
-    const ThreadIndent indent{this, std::this_thread::get_id()};
-    {
-        const scoped_read_lock_guard lock{mutex};
-        if (running.find(indent) != running.end()) [[unlikely]] {
-            return 0;
-        }
-    }
-
-    {
-        const scoped_write_lock_guard lock{mutex};
-        running.insert(indent);
-    }
-    try {
-        const ssize_t result = HashValueImpl();
-        {
-            const scoped_write_lock_guard lock{mutex};
-            running.erase(indent);
-        }
-        return result;
-    } catch (...) {
-        {
-            const scoped_write_lock_guard lock{mutex};
-            running.erase(indent);
-        }
-        std::rethrow_exception(std::current_exception());
-    }
-}
-
-py::object PyTreeSpec::ToPickleable() const {
-    const py::tuple node_states{GetNumNodes()};
-    ssize_t i = 0;
-    for (const auto& node : m_traversal) {
-        const scoped_critical_section2 cs{
-            node.custom != nullptr ? py::handle{node.custom->type.ptr()} : py::handle{},
-            node.node_data};
-        TupleSetItem(node_states,
-                     i++,
-                     py::make_tuple(py::int_(static_cast<ssize_t>(node.kind)),
-                                    py::int_(node.arity),
-                                    node.node_data ? node.node_data : py::none(),
-                                    node.node_entries ? node.node_entries : py::none(),
-                                    node.custom != nullptr ? node.custom->type : py::none(),
-                                    py::int_(node.num_leaves),
-                                    py::int_(node.num_nodes),
-                                    node.original_keys ? node.original_keys : py::none()));
-    }
-    return py::make_tuple(node_states, py::bool_(m_none_is_leaf), py::str(m_namespace));
-}
-
-// NOLINTBEGIN[cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers]
-// NOLINTNEXTLINE[readability-function-cognitive-complexity]
-/*static*/ std::unique_ptr<PyTreeSpec> PyTreeSpec::FromPickleable(const py::object& pickleable) {
-    const auto state = thread_safe_cast<py::tuple>(pickleable);
-    if (state.size() != 3) [[unlikely]] {
-        throw std::runtime_error("Malformed pickled PyTreeSpec.");
-    }
-    bool none_is_leaf = false;
-    std::string registry_namespace{};
-    auto out = std::make_unique<PyTreeSpec>();
-    out->m_none_is_leaf = none_is_leaf = thread_safe_cast<bool>(state[1]);
-    out->m_namespace = registry_namespace = thread_safe_cast<std::string>(state[2]);
-    const auto node_states = thread_safe_cast<py::tuple>(state[0]);
-    for (const auto& item : node_states) {
-        const auto t = thread_safe_cast<py::tuple>(item);
-        Node& node = out->m_traversal.emplace_back();
-        node.kind = static_cast<PyTreeKind>(thread_safe_cast<ssize_t>(t[0]));
-        if (t.size() != 7) [[unlikely]] {
-            if (t.size() == 8) [[likely]] {
-                if (t[7].is_none()) [[likely]] {
-                    if (node.kind == PyTreeKind::Dict || node.kind == PyTreeKind::DefaultDict)
-                        [[unlikely]] {
-                        throw std::runtime_error("Malformed pickled PyTreeSpec.");
-                    }
-                } else [[unlikely]] {
-                    if (node.kind == PyTreeKind::Dict || node.kind == PyTreeKind::DefaultDict)
-                        [[likely]] {
-                        node.original_keys = thread_safe_cast<py::list>(t[7]);
-                    } else [[unlikely]] {
-                        throw std::runtime_error("Malformed pickled PyTreeSpec.");
-                    }
-                }
-            } else [[unlikely]] {
-                throw std::runtime_error("Malformed pickled PyTreeSpec.");
-            }
-        }
-        node.arity = thread_safe_cast<ssize_t>(t[1]);
-        switch (node.kind) {
-            case PyTreeKind::Leaf:
-            case PyTreeKind::None:
-            case PyTreeKind::Tuple:
-            case PyTreeKind::List: {
-                if (!t[2].is_none()) [[unlikely]] {
-                    throw std::runtime_error("Malformed pickled PyTreeSpec.");
-                }
-                break;
-            }
-
-            case PyTreeKind::Dict:
-            case PyTreeKind::OrderedDict: {
-                node.node_data = thread_safe_cast<py::list>(t[2]);
-                break;
-            }
-
-            case PyTreeKind::NamedTuple:
-            case PyTreeKind::StructSequence: {
-                node.node_data = thread_safe_cast<py::type>(t[2]);
-                break;
-            }
-
-            case PyTreeKind::DefaultDict:
-            case PyTreeKind::Deque:
-            case PyTreeKind::Custom: {
-                node.node_data = t[2];
-                break;
-            }
-
-            default:
-                INTERNAL_ERROR();
-        }
-        if (node.kind == PyTreeKind::Custom) [[unlikely]] {  // NOLINT
-            if (!t[3].is_none()) [[unlikely]] {
-                node.node_entries = thread_safe_cast<py::tuple>(t[3]);
-            }
-            if (t[4].is_none()) [[unlikely]] {
-                node.custom = nullptr;
-            } else [[likely]] {
-                if (none_is_leaf) [[unlikely]] {
-                    node.custom =
-                        PyTreeTypeRegistry::Lookup<NONE_IS_LEAF>(t[4], registry_namespace);
-                } else [[likely]] {
-                    node.custom =
-                        PyTreeTypeRegistry::Lookup<NONE_IS_NODE>(t[4], registry_namespace);
-                }
-            }
-            if (node.custom == nullptr) [[unlikely]] {
-                std::ostringstream oss{};
-                oss << "Unknown custom type in pickled PyTreeSpec: " << PyRepr(t[4]);
-                if (!registry_namespace.empty()) [[likely]] {
-                    oss << " in namespace " << PyRepr(registry_namespace);
-                } else [[unlikely]] {
-                    oss << " in the global namespace";
-                }
-                oss << ".";
-                throw std::runtime_error(oss.str());
-            }
-        } else if (!t[3].is_none() || !t[4].is_none()) [[unlikely]] {
-            throw std::runtime_error("Malformed pickled PyTreeSpec.");
-        }
-        node.num_leaves = thread_safe_cast<ssize_t>(t[5]);
-        node.num_nodes = thread_safe_cast<ssize_t>(t[6]);
-    }
-    out->m_traversal.shrink_to_fit();
-    return out;
-}
-// NOLINTEND[cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers]
-
-size_t PyTreeSpec::ThreadIndentTypeHash::operator()(
-    const std::pair<const PyTreeSpec*, std::thread::id>& p) const {
-    size_t seed = 0;
-    HashCombine(seed, p.first);
-    HashCombine(seed, p.second);
-    return seed;
-}
-
-// NOLINTNEXTLINE[readability-function-cognitive-complexity]
-/*static*/ int PyTreeSpec::PyTpTraverse(PyObject* self_base, visitproc visit, void* arg) {
-#if PY_VERSION_HEX >= 0x03090000  // Python 3.9
-    Py_VISIT(Py_TYPE(self_base));
-#endif
-    auto* const instance = reinterpret_cast<py::detail::instance*>(self_base);
-    if (!instance->get_value_and_holder().holder_constructed()) [[unlikely]] {
-        // The holder is not constructed yet. Skip the traversal to avoid segfault.
-        return 0;
-    }
-    auto& self = thread_safe_cast<PyTreeSpec&>(py::handle{self_base});
-    for (const auto& node : self.m_traversal) {
-        Py_VISIT(node.node_data.ptr());
-        Py_VISIT(node.node_entries.ptr());
-        Py_VISIT(node.original_keys.ptr());
-    }
-    return 0;
-}
-
-// NOLINTNEXTLINE[readability-function-cognitive-complexity]
-/*static*/ int PyTreeIter::PyTpTraverse(PyObject* self_base, visitproc visit, void* arg) {
-#if PY_VERSION_HEX >= 0x03090000  // Python 3.9
-    Py_VISIT(Py_TYPE(self_base));
-#endif
-    auto* const instance = reinterpret_cast<py::detail::instance*>(self_base);
-    if (!instance->get_value_and_holder().holder_constructed()) [[unlikely]] {
-        // The holder is not constructed yet. Skip the traversal to avoid segfault.
-        return 0;
-    }
-    auto& self = thread_safe_cast<PyTreeIter&>(py::handle{self_base});
-    for (const auto& pair : self.m_agenda) {
-        Py_VISIT(pair.first.ptr());
-    }
-    Py_VISIT(self.m_root.ptr());
-    return 0;
 }
 
 }  // namespace optree
