+++ src/treespec/unflatten.cpp
@@ -21,8 +21,8 @@ limitations under the License.
 #include "include/critical_section.h"
 #include "include/exceptions.h"
 #include "include/registry.h"
+#include "include/stdutils.h"
 #include "include/treespec.h"
-#include "include/utils.h"
 
 namespace optree {
